# BUGFIX: Improve performance on find*Aggregates