The queries to determine the child or parent aggregates have been joined the hierarchy relation (hr) twice. Once for the parent and once for the children. But the hr does contain already both anchorpoints, so a second join is not needed.

Replay without this change:
```
Replaying events for projection "doctrineDbalContentGraph" of Content Repository "default" ...
 60440/60440 [============================] 100% 13 mins, 59 secs/13 mins, 59 secs 52.0 MiB
```

Replay with this change:
```
Replaying events for projection "doctrineDbalContentGraph" of Content Repository "default" ...
 60440/60440 [============================] 100%  4 mins, 13 secs/4 mins, 13 secs  52.0 MiB
```

Fixes: #5269