@dlubitz 2024-09-27
> Sadly there are no test for `findParentNodeAggregates` and maybe the other `find*Aggregates` so the ci passing might not be a reliable indicator. Id say we need tests here as well ._.

But they are tested within the behat tests.