@mhsdesign 2024-09-30
It seems we found the culprit. The reason for this obsolete join is history. Originally the nodename was stored in the hierarchy relation table. And with https://github.com/neos/neos-development-collection/pull/5018/ as stated:

> Node names are now stored in the contentgraph's node table

The node name is now in the node table.

There are actually more hidden places that can seemingly be optimised we found at least for places where a `join` was annotated being made for the legacy reason:

```
// we need to join with the hierarchy relation, because we need the node name.
```