@dlubitz 2024-10-01
This seems to be a bit more complex than expected. I need to check if the tests are covering all cases and if the changes still make sence.