@dlubitz 2024-10-02
I've splitted the changes. This PR now only contains the first change I did. This only affects the NodeAggregates query, which we can safely improve, as they don't join subtree tags.

The other changes are now in a draft PR https://github.com/neos/neos-development-collection/pull/5273. I need to find a way to test these queries, especially the resulting subtree tags. As it makes a difference for subtree tags if I use the parentanchorpoint or the childanchorpoint to join the node data.